backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
faiss-cpu
pydantic
python-multipart
pypdf
python-docx
pdf2image
pillow
pytesseract
networkx
langchain
python-dotenv
